SN-25 SN-48 SN-47  renamed invitations to group_invitations, moved all group related invitations/requests functions on handler and repository layer to  groupHandler.go and groupRepository.go -- needs to be checked if everything works expectedly; fixed group_invitations table
